Commit initial files
this is a Scrimba learning project.  Initially, it was a dynamic birthday card, however, now I have modified it as a Joe Cocker tribute card.  The GIF images have been taken from the internet, but as they are for personal use and not for publication I have not paid for them.